requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.8.0
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
cachetools==6.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.56.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.44.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.7
imageio==2.36.0
imgviz==1.7.6
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.11.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
labelme==5.7.0
lazy_loader==0.4
llvmlite==0.43.0
loguru==0.7.3
lxml==5.2.2
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
openai==2.4.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
osam==0.2.2
packaging==24.1
pandas==2.2.2
pillow==11.0.0
pillow_heif==0.20.0
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.28.3
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMatting==1.1.12
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rembg==2.0.59
requests==2.31.0
requests-mock==1.12.1
responses==0.25.0
rpds-py==0.20.1
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.45.3
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.0.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.1
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.0.2
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-mock==3.12.0
